feedback.md
Here is what our general feedback will look like until we get to react.  You will get a Pull Request like this where you can see the feedback in the changed files section, and then either choose to merge this file and keep it, or reject and close the pull request to delete it.